requirements.txt
# ── Core LangChain ────────────────────────────────────────────────────────────
langchain==0.3.25
langchain-community==0.3.24
langchain-google-genai==2.1.4

# ── Embeddings ─────────────────────────────────────────────────────────────────
sentence-transformers==3.4.1

# ── Vector Store ───────────────────────────────────────────────────────────────
faiss-cpu==1.10.0

# ── UI ─────────────────────────────────────────────────────────────────────────
streamlit==1.44.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.1.0